DEV TO RELEASE | Automation updates

+++ src/controllers/summarized-rangerdistrict-trapping.js
@@ -3,21 +3,64 @@ import {
   UnsummarizedTrappingModel,
 } from '../models';
 
-import { RESPONSE_TYPES } from '../constants';
+import {
+  RESPONSE_TYPES,
+  STATE_RANGER_DISTRICT_NAME_MAPPING,
+} from '../constants';
 
 import {
   aggregationPipelineCreator,
   cleanBodyCreator,
   csvDownloadCreator,
+  csvUploadCreator,
+  getIndexes,
   getModelAttributes,
   matchStateYear,
   newError,
+  upsertOpCreator,
 } from '../utils';
 
 const modelAttributes = getModelAttributes(SummarizedRangerDistrictTrappingModel);
 
 // this is a function to clean req.body
 const cleanBody = cleanBodyCreator(modelAttributes);
+
+const upsertOp = upsertOpCreator(getIndexes(SummarizedRangerDistrictTrappingModel));
+
+// function for cleaning row of csv (will cast undefined to null for specified fields)
+const cleanCsv = (row) => ({
+  ...row,
+  cleridCount: row.cleridCount ?? null,
+  cleridPerDay: row.cleridPerDay ?? {},
+  spbCount: row.spbCount ?? null,
+  spbPerDay: row.spbPerDay ?? {},
+  spots: null,
+  spotst1: null,
+  spotst2: null,
+  totalTrappingDays: row.totalTrappingDays ?? null,
+  trapCount: row.trapCount ?? null,
+});
+
+// transform ranger district name
+const rangerDistrictNameTransform = (row) => ({
+  ...row,
+  rangerDistrict: STATE_RANGER_DISTRICT_NAME_MAPPING[row.state]?.[row.rangerDistrict],
+});
+
+/**
+ * @description uploads a csv to the summarized ranger district collection
+ * @param {String} filename the csv filename on disk
+ * @throws RESPONSE_TYPES.BAD_REQUEST for missing fields
+ * @throws other errors depending on what went wrong
+ */
+export const uploadCsv = csvUploadCreator(
+  SummarizedRangerDistrictTrappingModel,
+  cleanCsv,
+  cleanBody,
+  undefined,
+  rangerDistrictNameTransform,
+  upsertOp,
+);
 
 /**
  * @description downloads a csv of the entire collection
@@ -94,10 +137,11 @@ export const deleteById = async (id) => {
 
 /**
  * @description Deletes all data in the collection
+ * @param {Object={}} [options] optional options object
  * @returns {Promise}
  */
-export const deleteAll = async () => {
-  return SummarizedRangerDistrictTrappingModel.deleteMany();
+export const deleteAll = async (options = {}) => {
+  return SummarizedRangerDistrictTrappingModel.deleteMany(options);
 };
 
 /**
@@ -107,6 +151,7 @@ export const deleteAll = async () => {
  * @returns {Promise}
  */
 export const deleteStateYear = async (state, year) => {
+  if (state === 2018) throw newError(RESPONSE_TYPES.BAD_REQUEST, 'Cannot delete 2018 data');
   return SummarizedRangerDistrictTrappingModel.deleteMany({ state, year });
 };
